refactor: :truck: change all components to components folder root
