+++ components/Hero/index.tsx
