+++ components/TypeWriter/index.tsx
