+++ components/TypeWriter/styles.module.css
