+++ pages/_app.tsx
@@ -2,7 +2,7 @@ import type { AppProps } from 'next/app'
 import { NextIntlProvider } from 'next-intl'
 
 import Layout from 'components/layout'
-import GA from 'components/global/GA'
+import GA from 'components/GA'
 import 'styles/globals.css'
 
 function MyApp({ Component, pageProps }: AppProps) {
