+++ pages/index.tsx
@@ -1,8 +1,8 @@
 import type { NextPage, GetStaticProps } from 'next'
 import { useTranslations } from 'next-intl'
 
-import Hero from 'components/global/Hero'
-import TypeWriter from 'components/global/TypeWriter'
+import Hero from 'components/Hero'
+import TypeWriter from 'components/TypeWriter'
 import Code from 'components/Icons/Code'
 
 const text2bold = {
@@ -14,13 +14,17 @@ const Home: NextPage = () => {
 
   return (
     <>
-      <Hero
-        greeting={t.rich('hero.greeting', text2bold)}
-        knowledge={<TypeWriter prefix={t('hero.knowledge.prefix')} titles={['Full Stack Developer', 'Web Developer', 'VTEX Developer']} />}
-        from={t.rich('hero.from', text2bold)}
-      >
-        <Code />
-      </Hero>
+      <section id="about">
+        <Hero
+          greeting={t.rich('hero.greeting', text2bold)}
+          knowledge={<TypeWriter prefix={t('hero.knowledge.prefix')} titles={['Full Stack Developer', 'Web Developer', 'VTEX Developer']} />}
+          from={t.rich('hero.from', text2bold)}
+        >
+          <Code />
+        </Hero>
+      </section>
+      <section id="projects"></section>
+      <section id="blog"></section>
     </>
   )
 }
